requirements.txt
# Core ML/AI dependencies
torch>=2.5.0
transformers>=4.51.0
accelerate>=0.34.0
peft>=0.12.0
bitsandbytes>=0.44.0
datasets>=3.0.0
tokenizers>=0.19.0
safetensors>=0.4.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0
datasets>=3.0.0

# Training utilities
wandb>=0.17.0
tensorboard>=2.15.0
tqdm>=4.66.0
rich>=13.6.0
typer>=0.9.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Code evaluation
evaluate>=0.4.0
bleu>=0.6.0
rouge-score>=0.1.0

# Speed optimization
flash-attn>=2.6.0
xformers>=0.0.23
einops>=0.7.0

# Memory optimization
deepspeed>=0.14.0
fsdp>=0.1.0

# Utilities
pyyaml>=6.0
jsonlines>=4.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocstrings[python]>=0.24.0

# Bend/HVM verification dependencies
# Note: Bend and HVM are installed via Rust/cargo, not pip
# See scripts/setup_bend_hvm.sh for installation
subprocess>=0.0.0  # Already in stdlib, but explicitly listed
tempfile>=0.0.0    # Already in stdlib, but explicitly listed
pathlib>=0.0.0     # Already in stdlib, but explicitly listed

# Optional: For extended functionality
# jupyter>=1.0.0  # For notebook support
# streamlit>=1.28.0  # For web interface
# fastapi>=0.104.0  # For API server
# gradio>=4.0.0  # For web UI